backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codegen-backend"
version = "0.1.0"
description = "FastAPI backend for code generation"
authors = [{ name = "Albert Anguera Sempere" }]
requires-python = "==3.12.*"
dependencies = [
  "fastapi==0.112.0",
  "uvicorn[standard]==0.30.3",
  "pydantic==2.8.2",
  "python-dotenv==1.0.1",
  "openai==1.40.6",
  "httpx==0.27.0",
  "orjson==3.10.7",
  "notebook>=7.4.5",
  "langchain-ollama>=0.2.1",
  "langchain>=0.3.27",
  "langchain-openai>=0.2.2",
  "langgraph>=0.6.5",
  "langchain-community>=0.3.27",
  "chromadb>=1.0.20",
  "sentence-transformers>=5.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
